# Faster path to check an IP address against known networks