… vendor/github.com/k-sone/critbitgo/net.go
… vendor/manifest